Move all helper methods for bilteralSwap and dividendComputation into proofUtils
Deleted helpers.js for bilteralSwap and dividendComputation after this. This is DRYifying the code - lots of duplication across the various helper files

Renamed .toBnAndAppendPoints() to .convertToBNAndAppendPoints() to make more descriptive
